fix: app crash on blank search API response for /records/
This prevents the app from crashing with an undefined check
(keeping the webserver alive to service other requests)
and adds a log message to hopefully tell us
what the problematic search API request is

I also noticed /mediarecords/ URLs might be vulnerable
to the same thing. I applied the same fix to that too.